Modify the database reset-version command
