+++ tools/database/src/commands.rs
@@ -9,9 +9,9 @@ use crate::corrupt::CorruptStateSnapshotCommand;
 use crate::drop_column::DropColumnCommand;
 use crate::make_snapshot::MakeSnapshotCommand;
 use crate::memtrie::LoadMemTrieCommand;
-use crate::reset_version::ResetVersionCommand;
 use crate::resharding_v2::ReshardingV2Command;
 use crate::run_migrations::RunMigrationsCommand;
+use crate::set_version::SetVersionCommand;
 use crate::state_perf::StatePerfCommand;
 use crate::write_to_db::WriteCryptoHashCommand;
 use clap::Parser;
@@ -67,8 +67,8 @@ enum SubCommand {
     /// Analyze size of contracts present in the current state
     AnalyzeContractSizes(AnalyzeContractSizesCommand),
 
-    /// Reset the database to the version used by the binary.
-    ResetVersion(ResetVersionCommand),
+    /// Manually set database version
+    SetVersion(SetVersionCommand),
 
     /// Perform on demand resharding V2
     Resharding(ReshardingV2Command),
@@ -98,7 +98,7 @@ impl DatabaseCommand {
             SubCommand::HighLoadStats(cmd) => cmd.run(home),
             SubCommand::AnalyzeDelayedReceipt(cmd) => cmd.run(home, genesis_validation),
             SubCommand::AnalyzeContractSizes(cmd) => cmd.run(home, genesis_validation),
-            SubCommand::ResetVersion(cmd) => cmd.run(home, genesis_validation),
+            SubCommand::SetVersion(cmd) => cmd.run(home, genesis_validation),
             SubCommand::Resharding(cmd) => {
                 let near_config = load_config(home, genesis_validation);
                 cmd.run(near_config, home)
