+++ tools/database/src/lib.rs
@@ -11,9 +11,9 @@ mod corrupt;
 mod drop_column;
 mod make_snapshot;
 mod memtrie;
-mod reset_version;
 mod resharding_v2;
 mod run_migrations;
+mod set_version;
 mod state_perf;
 mod utils;
 mod write_to_db;
